requirements-dev.txt
pytest
pytest-mock
requests-mock
pre-commit
black
isort
mypy